Switching to user ID

+++ src/User/User.cpp
@@ -26,6 +26,8 @@
 #include <DB/DB.h>
 #include <Crypto/Sha1String.h>
 #include <global.h>
+#include <WS/Events.h>
+#include <Logger/LoggerAPI.h>
 
 using namespace std;
 
@@ -35,20 +37,21 @@ User::User()
 {
 }
 
-User::User(DB *db,const string &user_name)
+User::User(DB *db,unsigned int user_id)
 {
-	db->QueryPrintf("SELECT user_login,user_password,user_profile,user_preferences FROM t_user WHERE user_login=%s",&user_name);
+	db->QueryPrintf("SELECT user_id, user_login,user_password,user_profile,user_preferences FROM t_user WHERE user_id=%i",&user_id);
 	
 	if(!db->FetchRow())
 		throw Exception("User","Unknown User");
 	
-	this->user_name = db->GetField(0);
-	user_password = db->GetField(1);
-	user_profile = db->GetField(2);
-	user_preferences = db->GetField(3);
+	this->user_id = db->GetFieldInt(0);
+	user_name = db->GetField(1);
+	user_password = db->GetField(2);
+	user_profile = db->GetField(3);
+	user_preferences = db->GetField(4);
 	
 	// Load rights
-	db->QueryPrintf("SELECT workflow_id, user_right_edit, user_right_read, user_right_exec, user_right_kill FROM t_user_right WHERE user_login=%s",&user_name);
+	db->QueryPrintf("SELECT workflow_id, user_right_edit, user_right_read, user_right_exec, user_right_kill FROM t_user_right WHERE user_id=%i",&user_id);
 	while(db->FetchRow())
 	{
 		unsigned int workflow_id = db->GetFieldInt(0);
@@ -116,11 +119,12 @@ bool User::CheckUserName(const string &user_name)
 	return true;
 }
 
-void User::Get(const string &name, QueryResponse *response)
+void User::Get(unsigned int id, QueryResponse *response)
 {
-	User user = Users::GetInstance()->Get(name);
+	User user = Users::GetInstance()->Get(id);
 	
 	DOMElement node = (DOMElement)response->AppendXML("<user />");
+	node.setAttribute("id",to_string(user.GetID()));
 	node.setAttribute("name",user.GetName());
 	node.setAttribute("profile",user.GetProfile());
 	node.setAttribute("preferences",user.GetPreferences());
@@ -142,6 +146,9 @@ unsigned int User::Create(const string &name, const string &password, const stri
 {
 	create_edit_check(name,password,profile);
 	
+	if(password=="")
+		throw Exception("User","Empty password","INVALID_PARAMETER");
+	
 	string password_sha1 = Sha1String(password).GetHex();
 	
 	DB db;
@@ -154,26 +161,34 @@ unsigned int User::Create(const string &name, const string &password, const stri
 	return db.InsertID();
 }
 
-void User::Edit(const string &name,const string &password, const string &profile)
+void User::Edit(unsigned int id, const std::string &name, const string &password, const string &profile)
 {
 	create_edit_check(name,password,profile);
 	
-	if(!Users::GetInstance()->Exists(name))
+	if(!Users::GetInstance()->Exists(id))
 		throw Exception("User","User not found","UNKNOWN_USER");
 	
-	string password_sha1 = Sha1String(password).GetHex();
-	
 	DB db;
-	db.QueryPrintf("UPDATE t_user SET user_password=%s,user_profile=%s WHERE user_login=%s",
-		&password_sha1,
+	db.QueryPrintf("UPDATE t_user SET user_login=%s, user_profile=%s WHERE user_id=%i",
+		&name,
 		&profile,
-		&name
+		&id
 	);
+	
+	if(password!="")
+	{
+		string password_sha1 = Sha1String(password).GetHex();
+		
+		db.QueryPrintf("UPDATE t_user SET user_password=%s WHERE user_id=%i",
+			&password_sha1,
+			&id
+		);
+	}
 }
 
-void User::ChangePassword(const string &name,const string &password)
+void User::ChangePassword(unsigned int id,const string &password)
 {
-	if(!Users::GetInstance()->Exists(name))
+	if(!Users::GetInstance()->Exists(id))
 		throw Exception("User","User not found","UNKNOWN_USER");
 	
 	if(password=="")
@@ -182,52 +197,52 @@ void User::ChangePassword(const string &name,const string &password)
 	string password_sha1 = Sha1String(password).GetHex();
 	
 	DB db;
-	db.QueryPrintf("UPDATE t_user SET user_password=%s WHERE user_login=%s",
+	db.QueryPrintf("UPDATE t_user SET user_password=%s WHERE user_id=%i",
 		&password_sha1,
-		&name
+		&id
 	);
 }
 
-void User::UpdatePreferences(const string &name, const string &preferences)
+void User::UpdatePreferences(unsigned int id, const string &preferences)
 {
-	if(!Users::GetInstance()->Exists(name))
+	if(!Users::GetInstance()->Exists(id))
 		throw Exception("User","User not found","UNKNOWN_USER");
 	
 	DB db;
-	db.QueryPrintf("UPDATE t_user SET user_preferences=%s WHERE user_login=%s",
+	db.QueryPrintf("UPDATE t_user SET user_preferences=%s WHERE user_id=%i",
 		&preferences,
-		&name
+		&id
 	);
 }
 
-void User::Delete(const string &name)
+void User::Delete(unsigned int id)
 {
 	DB db;
 	
 	db.StartTransaction();
 	
-	db.QueryPrintf("DELETE FROM t_user WHERE user_login=%s",&name);
+	db.QueryPrintf("DELETE FROM t_user WHERE user_id=%i",&id);
 	
 	if(db.AffectedRows()==0)
 		throw Exception("User","User not found","UNKNOWN_USER");
 	
-	db.QueryPrintf("DELETE FROM t_user_right WHERE user_login=%s",&name);
+	db.QueryPrintf("DELETE FROM t_user_right WHERE user_id=%i",&id);
 	
 	db.CommitTransaction();
 }
 
-void User::ClearRights(const string &name)
+void User::ClearRights(unsigned int id)
 {
-	if(!Users::GetInstance()->Exists(name))
+	if(!Users::GetInstance()->Exists(id))
 		throw Exception("User","User not found","UNKNOWN_USER");
 	
 	DB db;
-	db.QueryPrintf("DELETE FROM t_user_right WHERE user_login=%s",&name);
+	db.QueryPrintf("DELETE FROM t_user_right WHERE user_id=%i",&id);
 }
 
-void User::ListRights(const string &name, QueryResponse *response)
+void User::ListRights(unsigned int id, QueryResponse *response)
 {
-	User user = Users::GetInstance()->Get(name);
+	User user = Users::GetInstance()->Get(id);
 	
 	for(auto it=user.rights.begin();it!=user.rights.end();++it)
 	{
@@ -240,9 +255,9 @@ void User::ListRights(const string &name, QueryResponse *response)
 	}
 }
 
-void User::GrantRight(const string &name, unsigned int workflow_id, bool edit, bool read, bool exec, bool kill)
+void User::GrantRight(unsigned int id, unsigned int workflow_id, bool edit, bool read, bool exec, bool kill)
 {
-	if(!Users::GetInstance()->Exists(name))
+	if(!Users::GetInstance()->Exists(id))
 		throw Exception("User","User not found","UNKNOWN_USER");
 	
 	if(!Workflows::GetInstance()->Exists(workflow_id))
@@ -251,18 +266,18 @@ void User::GrantRight(const string &name, unsigned int workflow_id, bool edit, b
 	DB db;
 	int iedit = edit, iread = read, iexec = exec, ikill = kill;
 	db.QueryPrintf(
-		"INSERT INTO t_user_right(user_login, workflow_id, user_right_edit, user_right_read, user_right_exec, user_right_kill) VALUES(%s,%i,%i,%i,%i,%i)",
-		&name, &workflow_id, &iedit, &iread, &iexec, &ikill
+		"INSERT INTO t_user_right(user_id, workflow_id, user_right_edit, user_right_read, user_right_exec, user_right_kill) VALUES(%i,%i,%i,%i,%i,%i)",
+		&id, &workflow_id, &iedit, &iread, &iexec, &ikill
 	);
 }
 
-void User::RevokeRight(const string &name, unsigned int workflow_id)
+void User::RevokeRight(unsigned int id, unsigned int workflow_id)
 {
-	if(!Users::GetInstance()->Exists(name))
+	if(!Users::GetInstance()->Exists(id))
 		throw Exception("User","User not found","UNKNOWN_USER");
 	
 	DB db;
-	db.QueryPrintf("DELETE FROM t_user_right WHERE workflow_id=%i",&workflow_id);
+	db.QueryPrintf("DELETE FROM t_user_right WHERE user_id=%i AND workflow_id=%i",&id, &workflow_id);
 	
 	if(db.AffectedRows()==0)
 		throw Exception("User","No right found on that workflow","ALREADY_NO_RIGHTS");
@@ -275,9 +290,19 @@ void User::create_edit_check(const std::string &name, const std::string &passwor
 	
 	if(profile!="ADMIN" && profile!="USER")
 		throw Exception("User","Invalid profile, should be 'ADMIN' or 'USER'","INVALID_PARAMETER");
+}
+
+unsigned int get_id_from_query(XMLQuery *query)
+{
+	unsigned int id = query->GetRootAttributeInt("id", 0);
+	if(id!=0)
+		return id;
 	
-	if(password=="")
-		throw Exception("User","Empty password","INVALID_PARAMETER");
+	string name = query->GetRootAttribute("name","");
+	if(name=="")
+		throw Exception("User","Missing 'id' or 'name' attribute","MISSING_PARAMETER");
+	
+	return Users::GetInstance()->Get(name).GetID();
 }
 
 bool User::HandleQuery(const User &user, XMLQuery *query, QueryResponse *response)
@@ -286,19 +311,19 @@ bool User::HandleQuery(const User &user, XMLQuery *query, QueryResponse *respons
 	
 	if(action=="get")
 	{
-		string name = query->GetRootAttribute("name");
+		unsigned int id = get_id_from_query(query);
 		
-		if(!user.IsAdmin() && user.GetName()!=name)
+		if(!user.IsAdmin() && user.GetID()!=id)
 			User::InsufficientRights();
 		
-		Get(name,response);
+		Get(id,response);
 		
 		return true;
 	}
 	else if(action=="create" || action=="edit")
 	{
 		string name = query->GetRootAttribute("name");
-		string password = query->GetRootAttribute("password");
+		string password = query->GetRootAttribute("password", "");
 		string profile = query->GetRootAttribute("profile");
 		
 		if(action=="create")
@@ -308,13 +333,23 @@ bool User::HandleQuery(const User &user, XMLQuery *query, QueryResponse *respons
 			
 			unsigned int id = Create(name, password, profile);
 			response->GetDOM()->getDocumentElement().setAttribute("user-id",to_string(id));
+			
+			LoggerAPI::LogAction(user,id,"User",query->GetQueryGroup(),action);
+			
+			Events::GetInstance()->Create(Events::en_types::USER_CREATED, id);
 		}
 		else
 		{
 			if(!user.IsAdmin())
 				User::InsufficientRights();
 			
-			Edit(name, password, profile);
+			unsigned int id = get_id_from_query(query);
+			
+			Edit(id, name, password, profile);
+			
+			LoggerAPI::LogAction(user,id,"User",query->GetQueryGroup(),action);
+			
+			Events::GetInstance()->Create(Events::en_types::USER_MODIFIED, id);
 		}
 		
 		Users::GetInstance()->Reload();
@@ -323,29 +358,38 @@ bool User::HandleQuery(const User &user, XMLQuery *query, QueryResponse *respons
 	}
 	else if(action=="change_password")
 	{
-		string name = query->GetRootAttribute("name");
+		unsigned int id = get_id_from_query(query);
+		
 		string password = query->GetRootAttribute("password");
 		
-		if(!user.IsAdmin() && user.GetName()!=name)
+		if(!user.IsAdmin() && user.GetID()!=id)
 			User::InsufficientRights();
 		
-		ChangePassword(name,password);
+		ChangePassword(id,password);
 		
 		Users::GetInstance()->Reload();
+		
+		LoggerAPI::LogAction(user,id,"User",query->GetQueryGroup(),action);
+		
+		Events::GetInstance()->Create(Events::en_types::USER_MODIFIED, id);
 		
 		return true;
 	}
 	else if(action=="update_preferences")
 	{
-		string name = query->GetRootAttribute("name");
+		unsigned int id = get_id_from_query(query);
 		string preferences = query->GetRootAttribute("preferences");
 		
-		if(!user.IsAdmin() && user.GetName()!=name)
+		if(!user.IsAdmin() && user.GetID()!=id)
 			User::InsufficientRights();
 		
-		UpdatePreferences(name,preferences);
+		UpdatePreferences(id,preferences);
 		
 		Users::GetInstance()->Reload();
+		
+		LoggerAPI::LogAction(user,id,"User",query->GetQueryGroup(),action);
+		
+		Events::GetInstance()->Create(Events::en_types::USER_MODIFIED, id);
 		
 		return true;
 	}
@@ -354,11 +398,15 @@ bool User::HandleQuery(const User &user, XMLQuery *query, QueryResponse *respons
 		if(!user.IsAdmin())
 			User::InsufficientRights();
 		
-		string name = query->GetRootAttribute("name");
+		unsigned int id = get_id_from_query(query);
 		
-		Delete(name);
+		Delete(id);
 		
 		Users::GetInstance()->Reload();
+		
+		LoggerAPI::LogAction(user,id,"User",query->GetQueryGroup(),action);
+		
+		Events::GetInstance()->Create(Events::en_types::USER_REMOVED, id);
 		
 		return true;
 	}
@@ -367,11 +415,15 @@ bool User::HandleQuery(const User &user, XMLQuery *query, QueryResponse *respons
 		if(!user.IsAdmin())
 			User::InsufficientRights();
 		
-		string name = query->GetRootAttribute("name");
+		unsigned int id = get_id_from_query(query);
 		
-		ClearRights(name);
+		ClearRights(id);
 		
 		Users::GetInstance()->Reload();
+		
+		LoggerAPI::LogAction(user,id,"User",query->GetQueryGroup(),action);
+		
+		Events::GetInstance()->Create(Events::en_types::USER_MODIFIED, id);
 		
 		return true;
 	}
@@ -380,9 +432,13 @@ bool User::HandleQuery(const User &user, XMLQuery *query, QueryResponse *respons
 		if(!user.IsAdmin())
 			User::InsufficientRights();
 		
-		string name = query->GetRootAttribute("name");
+		unsigned int id = get_id_from_query(query);
 		
-		ListRights(name, response);
+		ListRights(id, response);
+		
+		LoggerAPI::LogAction(user,id,"User",query->GetQueryGroup(),action);
+		
+		Events::GetInstance()->Create(Events::en_types::USER_MODIFIED, id);
 		
 		return true;
 	}
@@ -391,16 +447,20 @@ bool User::HandleQuery(const User &user, XMLQuery *query, QueryResponse *respons
 		if(!user.IsAdmin())
 			User::InsufficientRights();
 		
-		string name = query->GetRootAttribute("name");
+		unsigned int id = get_id_from_query(query);
 		unsigned int workflow_id = query->GetRootAttributeInt("workflow_id");
 		bool edit = query->GetRootAttributeBool("edit");
 		bool read = query->GetRootAttributeBool("read");
 		bool exec = query->GetRootAttributeBool("exec");
 		bool kill = query->GetRootAttributeBool("kill");
 		
-		GrantRight(name, workflow_id, edit, read, exec, kill);
+		GrantRight(id, workflow_id, edit, read, exec, kill);
 		
 		Users::GetInstance()->Reload();
+		
+		LoggerAPI::LogAction(user,id,"User",query->GetQueryGroup(),action);
+		
+		Events::GetInstance()->Create(Events::en_types::USER_MODIFIED, id);
 		
 		return true;
 	}
@@ -409,12 +469,16 @@ bool User::HandleQuery(const User &user, XMLQuery *query, QueryResponse *respons
 		if(!user.IsAdmin())
 			User::InsufficientRights();
 		
-		string name = query->GetRootAttribute("name");
+		unsigned int id = get_id_from_query(query);
 		unsigned int workflow_id = query->GetRootAttributeInt("workflow_id");
 		
-		RevokeRight(name, workflow_id);
+		RevokeRight(id, workflow_id);
 		
 		Users::GetInstance()->Reload();
+		
+		LoggerAPI::LogAction(user,id,"User",query->GetQueryGroup(),action);
+		
+		Events::GetInstance()->Create(Events::en_types::USER_MODIFIED, id);
 		
 		return true;
 	}
